+++ examples/UploadCandidatesFromSample.py
@@ -5,11 +5,10 @@ from functools import partial
 from multiprocessing import Pool
 
 from skills_ml.algorithms.sampling import Sample
-from skills_ml.algorithms.skill_extractors.freetext import \
+from skills_ml.algorithms.skill_extractors import \
     FuzzyMatchSkillExtractor,\
     ExactMatchSkillExtractor,\
-    OccupationScopedSkillExtractor
-from skills_ml.algorithms.skill_extractors import upload_candidates_from_job_posting_json
+    upload_candidates_from_job_posting_json
 
 
 def generate_skill_candidates_multiprocess(candidates_path, sample, skill_extractor, n_jobs):
