+++ skills_ml/algorithms/skill_list_processors/__init__.py
@@ -1,0 +1,3 @@
+from .onet_ksat import OnetSkillListProcessor
+
+__all__ = ('OnetSkillListProcessor',)
